refactor listdeletedcommand to recyclebincommand
