+++ src/main/java/seedu/address/logic/parser/RecycleBinParser.java
@@ -4,8 +4,7 @@ import seedu.address.logic.commands.ClearCommand;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.DeleteCommand;
 import seedu.address.logic.commands.HelpCommand;
-import seedu.address.logic.commands.ListCommand;
-import seedu.address.logic.commands.ListDeletedCommand;
+import seedu.address.logic.commands.RecycleBinCommand;
 import seedu.address.logic.commands.RestoreCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
@@ -28,8 +27,8 @@ public class RecycleBinParser extends BasicParser {
 
         switch (commandWord) {
 
-            case ListDeletedCommand.COMMAND_WORD:
-                return new ListDeletedCommand();
+            case RecycleBinCommand.COMMAND_WORD:
+                return new RecycleBinCommand();
 
             case DeleteCommand.COMMAND_WORD:
                 return new DeleteCommandParser().parse(arguments);
